Add 'sort' parameter to browse(), search() and query() functions
'sort' is a list of metadata keys.

Sources will try to sort the results following the list of keys. If some of the
keys are not supported or can be used to sort the results, then they will be
skipped. As case extreme, if source does not support sorting, it will ignore
this parameter.