core: Free specs after last element is sent
So far, specs were inmediatly destroyed when sources invoke callback with
remaining == 0.

This has the drawback that after it, spec is no longer valid, and sources
should avoid to use them. This means that sources must copy spec content if
they want to use them after the callback.

To avoid this problem, add an idle function to free the spec after source
invokes callback with remaining == 0. So source is safe to continue accessing
the spec content, and when it returns, the idle function to free the spec will
be activated.

Even more, if source, for any reason, needs to keep the spec even after
returning (could be the spec is in turn wrapped in another structure handled in
an idle function), it can ref() the spec, so it is not destroyed in the return,
but later, when source calls the appropriate unref() function.