core: Free helper structures after processing metadata
Both in grl_metadata_source_resolve() and grl_metadata_source_set_metadata()
some helper structures are used, being GrlMetadataSourceFooSpec ones the most
important.

When source finishes the task, it calls a callback. This patch enqueues a task
that will free both the helpers structures and unref those specs after source's
caller finalishes.

It is up to source to ref() specs if it wants to keep them after finishing the
task.